src/python/training/requirements.txt
coloredlogs==15.0
decorator==4.4.2
numpy==1.18.5
pandas==1.2.3
Pillow==8.2.0
scikit-image==0.18.1
scipy==1.4.1
sklearn==0.0
tensorflow-gpu==2.3.0
tensorflow-probability==0.11.1
